Convert binary data to MapSet data independently
The previous version used BitFieldSet to create the test data, which
was kinda risky as the `.to_list/1` function on BitFieldSet could be
wrong. This implements an algorithm for converting the random binary to
a bit field stored in a MapSet, and that is used to check the
correctness of the implementation.

It is slow as a long winter, but at least it is independent from the
implementation we are checking against.