Use Brian Kernighan's Algorithm for getting the number of "on"-bits
Previously I iterated over all the bits in the set and incremented a
counter for each bit that was set. Of course this always resulted in
a number of iterations equal to the size of the bit field.

I have found a better way of doing this invented by Brian Kernighan;
by subtracting one from the current number and performing a binary AND
on that and the current number we can "pop off" the least significant
bit; now it is just a matter of counting how many times we can do that
and we will get the number of "on" bits in the bit field.

Example

``` elixir
iex> use Bitwise
Bitwise
iex> n = 0b1001
9
iex> n = band(n, n - 1)
8
iex> Integer.to_string(n, 2)
"1000"
iex> n = band(n, n - 1)
0
```

The number of on bits for the number 9 is 2. The previous
implementation would require 4 iterations, where the new one requires
2 to get to the same result. Though we are doing some comparisons this
should be considerably faster than the previous version.