Optimize the to_list/1 function for speed
Previously we iterated each bit in the bit-field and accumulated the
positions of the on bits. The new implementation will calculate the
next on bit with bitwise operations and math, so we can skip large
portions of the bit-field that is not set.

This implementation should perform considerably better for larger
bit-fields. This is important as the to_list/1 function is used to
display using the inspect protocol.