Replace the brute force difference function with a bitwise one
Instead of iterating both the sets when comparing for elements present
in a but not in b we can perform the same thing with three bitwise
operators; first we compare what is in b and a, then we reverse that
and use that to compare the content of a.

This implementation uses no conditional branching and comparisons
besides the bitwise operators.